+++ blocklets/ai-studio/src/components/file-editor/execute-block.tsx
@@ -208,7 +208,7 @@ export default function ExecuteBlockForm({
                       doc.transact(() => {
                         if (value.tools) {
                           delete value.tools[tool.id];
-                          Object.values(value.tools).forEach((i, index) => (i.index = index));
+                          sortBy(Object.values(value.tools), 'index').forEach((i, index) => (i.index = index));
                         }
                       });
                     }}>
@@ -296,6 +296,7 @@ export default function ExecuteBlockForm({
       </Stack>
 
       <ToolDialog
+        executeBlock={value}
         ref={toolForm}
         projectId={projectId}
         assistant={assistant}
@@ -305,10 +306,15 @@ export default function ExecuteBlockForm({
           const doc = (getYjsValue(value) as Map<any>).doc!;
           doc.transact(() => {
             value.tools ??= {};
+
+            const old = value.tools[tool.id];
+
             value.tools[tool.id] = {
-              index: Math.max(-1, ...Object.values(value.tools).map((i) => i.index)) + 1,
+              index: old?.index ?? Math.max(-1, ...Object.values(value.tools).map((i) => i.index)) + 1,
               data: tool,
             };
+
+            sortBy(Object.values(value.tools), 'index').forEach((tool, index) => (tool.index = index));
           });
           dialogState.close();
         }}
@@ -332,13 +338,14 @@ interface ToolDialogImperative {
 const ToolDialog = forwardRef<
   ToolDialogImperative,
   {
+    executeBlock: ExecuteBlockYjs;
     projectId: string;
     gitRef: string;
     onSubmit: (value: ToolDialogForm) => any;
     DialogProps?: DialogProps;
     assistant: AssistantYjs;
   }
->(({ assistant, projectId, gitRef, onSubmit, DialogProps }, ref) => {
+>(({ executeBlock, assistant, projectId, gitRef, onSubmit, DialogProps }, ref) => {
   const { t } = useLocaleContext();
   const { store } = useProjectStore(projectId, gitRef);
   const assistantId = assistant.id;
@@ -362,6 +369,15 @@ const ToolDialog = forwardRef<
     sortBy(Object.values(file.parameters), (i) => i.index).filter(
       (i): i is typeof i & { data: { key: string } } => !!i.data.key
     );
+
+  const assistantParameters = new Set([
+    ...Object.values(assistant.parameters ?? {}).map((i) => i.data.key),
+    ...(assistant.type === 'prompt'
+      ? Object.values(assistant.prompts ?? {})
+          .map((i) => (i.data.type === 'executeBlock' ? i.data.data.variable : undefined))
+          .filter(Boolean)
+      : []),
+  ]);
 
   const createFile = useCreateFile();
 
@@ -493,7 +509,13 @@ const ToolDialog = forwardRef<
                   name={`parameters.${parameter.key}`}
                   render={({ field }) => (
                     <PromptEditorField
-                      placeholder={`{{ ${parameter.key} }}`}
+                      placeholder={
+                        executeBlock.selectType === 'selectByPrompt'
+                          ? t('selectByPromptParameterPlaceholder')
+                          : assistantParameters.has(parameter.key)
+                          ? `{{ ${parameter.key} }}`
+                          : undefined
+                      }
                       value={field.value || ''}
                       projectId={projectId}
                       gitRef={gitRef}